[Add] 컬렉션뷰로 수정 (#48)
